refactor: 모달 내부 컨텐츠 모듈화
